+++ scripts/release.sh
@@ -1,0 +1,5 @@
+# Build Package
+yarn build
+
+# Publish
+yarn changeset publish